Fix the logout bug
