+++ package.json
@@ -59,6 +59,7 @@
     "ember-concurrency": "^1.1",
     "ember-crumbly": "^3.0.1",
     "ember-data": "~3.10.0",
+    "ember-fetch": "^7.0.1",
     "ember-group-by": "^0.0.6",
     "ember-intl": "^4.2.2",
     "ember-keyboard-shortcuts": "^1.2",
@@ -74,7 +75,7 @@
     "ember-resolver": "^7.0",
     "ember-responsive": "3.0.5",
     "ember-router-scroll": "^1.3.3",
-    "ember-simple-auth": "^2.1.1",
+    "ember-simple-auth": "^3.0",
     "ember-sortable": "^1.10.0",
     "ember-source": "~3.10.0",
     "ember-test-selectors": "^3.0",
